Starting refactor
