--- src/Message.tsx
@@ -1,5 +1,0 @@
-function Message() {
-    const name = "Jack";
-    return <h1>Hello {name}</h1>;
-}
-export default Message;
