+++ src/components/typingtest/BootstrapModal.tsx
