+++ src/components/typingtest/TextInput.tsx
