+++ src/components/typingtest/WordBox.tsx
